+++ src/client/_deploySlashCommands.ts
@@ -1,45 +1,28 @@
+// external libs
 import { REST } from '@discordjs/rest';
 import { Routes } from 'discord-api-types/v9';
 
-import config from './config';
+// lib & config
+import config from '../config/config';
+import { AllSlashCommands } from '../commands';
+import { devConsoleMessage } from '../messages';
+
+// interfaces
 import type { CombinedCommandsInt } from '../interfaces';
-import {
-	que,
-	status,
-	clearLast,
-	clearPost,
-	setCronfig,
-	currentCronfig,
-	startQueRoutine,
-	stopQueRoutine,
-	pauseQueRoutine,
-} from '../commands';
 
 const { token, client_id, guild } = config;
 const rest = new REST({ version: '9' }).setToken(token);
 
-let AllSlashCommands = [
-	que,
-	status,
-	clearLast,
-	clearPost,
-	setCronfig,
-	currentCronfig,
-	startQueRoutine,
-	stopQueRoutine,
-	pauseQueRoutine,
-];
-
 export const deploySlashCommands = async (): Promise<CombinedCommandsInt[]> => {
 	try {
-		console.log(`Updating Quebert's slash (/) commands...`);
+		devConsoleMessage(`Updating Quebert's slash (/) commands...`);
 		await rest.put(Routes.applicationGuildCommands(client_id!, guild!), {
 			body: AllSlashCommands.map((c) => c.data.toJSON()),
 		});
-		console.log(`Commands successfully updated.`);
+		devConsoleMessage(`Commands successfully updated.`);
 		return AllSlashCommands;
 	} catch (error) {
-		console.log(error);
+		devConsoleMessage(error);
 	}
 	return AllSlashCommands;
 };
